Update readme to point to V1 of Tailwinds
Tailwinds has now released v2. Upgrade this repo to Tailwinds v2 isn't really necessary, but this readme points to the default Tailwinds docs, which is for V2. Things like customising colours has now changed, and took me some time to realise I was reading the wrong docs for this repo.